Minor test fixes

+++ src/Cassandra.IntegrationTests/Core/ExceptionsTests.cs
@@ -23,7 +23,7 @@ using System.Threading;
 
 namespace Cassandra.IntegrationTests.Core
 {
-    [TestFixture, Category("short")]
+    [TestFixture, Category("long")]
     public class ExceptionsTests
     {
         public ExceptionsTests()
